==> Media+Publication/SDSHL/SDSHL/notebooks/2-experiments/0-gensim.ipynb <==
#paths required
datafolder_p = r'D:\18-DS\github\SDSHL\data\processed'
datafolder_e = r'D:\18-DS\github\SDSHL\data\external'
datafolder_i = r'D:\18-DS\github\SDSHL\data\internal'
modelfolder  = r'D:\18-DS\data\models'

vectorfolder_bert = modelfolder + r'\BERT'
vectorfolder_ft   = modelfolder + r'\fasttext_wiki.hi'

train_datafile =datafolder_p+r'\2-train.csv'
test_datafile =datafolder_p+r'\2-test.csv'

#from __future__ import print_function

import pandas as pd
import nltk
import numpy as np
import re
from sklearn.model_selection import train_test_split   

# Set values for various parameters
feature_size = 15    # Word vector dimensionality  
window_context = 20  # Context window size                                                                                    
min_word_count = 1   # Minimum word count                        
sample = 1e-3        # Downsample setting for frequent words
sg = 1               # skip-gram model

df =pd.read_csv(train_datafile, sep="\t")
corpus = list(df.sentence)
tokenized_corpus = [nltk.word_tokenize(doc) for doc in corpus]
df['sentence'] = tokenized_corpus

# np.random.seed(100)
# df_len = len(df)
# train_len = int(df_len*.75)
# train_idx = np.random.choice(range(df_len), train_len, replace=False)
# test_idx = [i for i in range(0,df_len) if i not in train_idx]
# train_idx.sort()
# test_idx.sort()

# fastText Embedding

### Transfer Learning. Already done in embedding_fastText so not doing again using Gensim

#Load vector as a model. With gensim framework

#from gensim.models import KeyedVectors

# Creating the model
## Takes a lot of time depending on the vector file size 

#ft_model = KeyedVectors.load_word2vec_format(vectorfolder_ft+'\wiki.hi.300.vec')


### Create a fastText based model

from gensim.models.fasttext import FastText
ft_model = FastText(tokenized_corpus, size=feature_size, 
                          window=window_context, min_count = min_word_count,
                          sg=sg, sample=sample, iter=5000)

ft_model.save( modelfolder + "\model_ft.bin")

### Create sentence embedding & save vector

hi_model.n_similarity('हम','मैं')

#hi_model.wv.word_vec('हम')

# hi_model.word_vec('हम')
# hi_model.wv['हम']

# GloVe Embedding

### Glove Vector is available for English Language so ignoring this

# #Load vector as a model. With gensim framework

# from gensim.models import KeyedVectors

# # Creating the model
# ## Takes a lot of time depending on the vector file size 
# # English GloVe
# glove_model = KeyedVectors.load_word2vec_format(vectorfolder_glove+'\glove.6B.50d.txt')

# Word2Vector Embedding

#from gensim.test.utils import common_texts, get_tmpfile
from gensim.models import Word2Vec

#path = get_tmpfile("word2vec.model")

w2v_model = Word2Vec(tokenized_corpus, size=100, window=5, min_count=1, workers=4)
w2v_model.save( modelfolder + "\model_w2v.bin")

### Create sentence embedding and save
