==> Media+Publication/SDSHL/SDSHL/notebooks/2-experiments/0-nltk.ipynb <==
import nltk
from nltk.tag import tnt
from nltk.corpus import indian
nltk.download('indian')

from nltk.tag import tnt
import nltk
from nltk.corpus import indian
train_data = indian.tagged_sents('hindi.pos')[0:530]
test_data = indian.tagged_sents('hindi.pos')[530:]
tnt_pos_tagger = tnt.TnT()
tnt_pos_tagger.train(train_data)

# evaluating 
a = tnt_pos_tagger.evaluate(test_data) 
  
print ("Accuracy of TnT Tagging : ", a) 

doc = """केंद्र की मोदी सरकार ने शुक्रवार को अपना अंतरिम बजट पेश किया. कार्यवाहक वित्त मंत्री पीयूष गोयल \
ने अपने बजट में किसान, मजदूर, करदाता, महिला वर्ग समेत हर किसी के लिए बंपर ऐलान किए. हालांकि, \
बजट के बाद भी टैक्स को लेकर काफी कन्फ्यूजन बना रहा. \
केंद्र सरकार के इस अंतरिम बजट क्या खास रहा और किसको क्या मिला, आसान भाषा में यहां समझें"""


tagged_words = (tnt_pos_tagger.tag(nltk.word_tokenize(doc)))
print(tagged_words)

train_data = indian.tagged_sents('hindi.pos')
tnt_pos_tagger = tnt.TnT()
tnt_pos_tagger.train(train_data)

text='जब तन्हा रातो में मोगैम्बो चिलाता है की मोगैम्बो खुश हुवा तो समझिये की मनमोहन सरकार के बुरे सपने याद आ गये,'
tagged_words = (tnt_pos_tagger.tag(nltk.word_tokenize(text)))
print(tagged_words)

# hindi.pos gives very bad result for pos

nltk.download('panlex_swadesh')
nltk.corpus.CorpusReader.fileids #.tagged_words()

#below is general command. But if we have corpus of tha langauge then only it will be useful.
#nltk.corpus.<corpus>.tagged_words()

#for english langugage 107 pos and other tag corpus available at http://www.nltk.org/nltk_data/

text = nltk.word_tokenize("And now for something completely different")
text

nltk.download('averaged_perceptron_tagger')

nltk.pos_tag(text)

import codecs

f = codecs.open(datafolder_p+"\Hinglish_Sarcasm_FE.txt",mode="r",encoding="utf-8")
f2 = codecs.open(datafolder_p+"\Hinglish_Sarcasm_FE2.txt",mode="w",encoding="utf-8")
cnt = 0
red = 0
for i in f.readlines():
	cnt += 1
	if ('|UNK' in i) or ('B-NP' in i) or ('I-NP' in i) or ('|O' in i):
		red += 1
	else:
		f2.write(i)
# print red
# print cnt
# print cnt-red

f.close()
f2.close()

import nltk
from nltk.corpus import indian
from nltk.tag import tnt
import string


nltk.download('punkt')
nltk.download()


tagged_set = 'hindi.pos'
word_set = indian.sents(tagged_set)
count = 0
for sen in word_set:
    count = count + 1
    sen = "".join([" "+i if not i.startswith("'") and i not in string.punctuation else i for i in sen]).strip()
    print (count, sen)
print ('Total sentences in the tagged file are',count)

train_perc = .9

train_rows = int(train_perc*count)
test_rows = train_rows + 1

print ('Sentences to be trained',train_rows, 'Sentences to be tested against',test_rows)


# In[ ]:


data = indian.tagged_sents(tagged_set)
train_data = data[:train_rows]
test_data = data[test_rows:]


pos_tagger = tnt.TnT()
pos_tagger.train(train_data)
pos_tagger.evaluate(test_data)


# In[ ]:



sentence_to_be_tagged = "३९ गेंदों में दो चौकों और एक छक्के की मदद से ३४ रन बनाने वाले परोरे अंत तक आउट नहीं हुए ।"

tokenized = nltk.word_tokenize(sentence_to_be_tagged)


print(pos_tagger.tag(tokenized))


# In[ ]:


data.df

nltk.download('treebank')

from nltk.tag import tnt 
from nltk.corpus import treebank 
  
# initializing training and testing set     
train_data = treebank.tagged_sents()[:3000] 
test_data = treebank.tagged_sents()[3000:] 
  
# initializing tagger 
tnt_tagging = tnt.TnT() 
  
# training 
tnt_tagging.train(train_data) 
  
# evaluating 
a = tnt_tagging.evaluate(test_data) 
  
print ("Accuracy of TnT Tagging : ", a) 

doc ='''Memory networks with an architecture similar to (Mahata et al., 2018), are the deep
learning models that were trained. A random classifier that randomly generated predictions from a label distribution similar to that
of the training dataset was also implemented.'''

tagged_words = (tnt_tagging.tag(nltk.word_tokenize(doc)))
print(tagged_words)
