==> Media+Publication/SDSHL/SDSHL/notebooks/2-experiments/0-standfordnlp.ipynb <==
#paths required
datafolder_p = r'D:\18-DS\github\SDSHL\data\processed'
datafolder_e = r'D:\18-DS\github\SDSHL\data\external'
datafolder_i = r'D:\18-DS\github\SDSHL\data\internal'
modelfolder  = r'D:\18-DS\data\models'

vectorfolder_bert = modelfolder + r'\BERT'
vectorfolder_ft   = modelfolder + r'\fasttext_wiki.hi'
vectorfolder_standford = modelfolder + r'\POS-Tagger-Hindi'

train_datafile =datafolder_p+r'\2-train.csv'
test_datafile =datafolder_p+r'\2-test.csv'

file =datafolder_p+ r'\2-Hinglish_Sarcasm_Clean.csv'

import stanfordnlp

MODELS_DIR = '.'
stanfordnlp.download('hi', MODELS_DIR) # Download the English models

nlp = stanfordnlp.Pipeline(processors='tokenize,pos', models_dir=folder, treebank='hi_hdtb', use_gpu=True, pos_batch_size=3000) # Build the pipeline, specify part-of-speech processor's batch size
doc = nlp("""केंद्र की मोदी सरकार ने शुक्रवार को अपना अंतरिम बजट पेश किया. कार्यवाहक वित्त मंत्री पीयूष गोयल \
ने अपने बजट में किसान, मजदूर, करदाता, महिला वर्ग समेत हर किसी के लिए बंपर ऐलान किए. हालांकि, \
बजट के बाद भी टैक्स को लेकर काफी कन्फ्यूजन बना रहा. \
केंद्र सरकार के इस अंतरिम बजट क्या खास रहा और किसको क्या मिला, आसान भाषा में यहां समझें""")
doc.sentences[0].print_tokens() # Look at the result

nlp = stanfordnlp.Pipeline(processors = "tokenize,pos", lang="hi", models_dir=folder, use_gpu=False)

import stanfordnlp

#tag = r'D:\18-DS\github\0githubBranched\hindi-pos-tagger\HMM\Code\hi_hdtb_models\hi_hdtb_tagger.pt'
#pret = r'D:\18-DS\github\0githubBranched\hindi-pos-tagger\HMM\Code\hi_hdtb_models\hi_hdtb.pretrain.pt'


tag = folder+r'\hi_hdtb_models\hi_hdtb_tagger.pt'
pret = folder+r'\hi_hdtb_models\hi_hdtb.pretrain.pt'


#'pos_model_path': './en_ewt_models/en_ewt_tagger.pt',
#'pos_pretrain_path': './en_ewt_models/en_ewt.pretrain.pt',
            

config = {
        'processors': 'tokenize,pos',
        'tokenize_pretokenized': True,
        'pos_model_path': tag,
        'pos_pretrain_path': pret,
        'pos_batch_size': 1000,
        'lang':'hi'
         }
nlp = stanfordnlp.Pipeline(**config)
doc = nlp("""केंद्र की मोदी सरकार ने शुक्रवार को अपना अंतरिम बजट पेश किया. कार्यवाहक वित्त मंत्री पीयूष गोयल \
ने अपने बजट में किसान, मजदूर, करदाता, महिला वर्ग समेत हर किसी के लिए बंपर ऐलान किए. हालांकि, \
बजट के बाद भी टैक्स को लेकर काफी कन्फ्यूजन बना रहा. \
केंद्र सरकार के इस अंतरिम बजट क्या खास रहा और किसको क्या मिला, आसान भाषा में यहां समझें""")
print(doc.conll_file.conll_as_string())

import stanfordnlp

nlp = stanfordnlp.Pipeline()
doc = nlp("Barack Obama was born in Hawaii.")
print(*[f'text: {word.text+" "}\tlemma: {word.lemma}\tupos: \
        {word.upos}\txpos: {word.xpos}' for sent in doc.sentences for word in sent.words], sep='\n')
