==> Media+Publication/SDSHL/SDSHL/notebooks/2-experiments/03-Transliteration.ipynb <==
#paths required
datafolder_p = r'D:\18-DS\github\SDSHL\data\processed'
datafolder_e = r'D:\18-DS\github\SDSHL\data\external'
datafolder_i = r'D:\18-DS\github\SDSHL\data\internal'
modelfolder  = r'D:\18-DS\data\models'

vectorfolder_bert = modelfolder + r'\BERT'
vectorfolder_ft   = modelfolder + r'\fasttext_wiki.hi'
vectorfolder_standford = modelfolder + r'\POS-Tagger-Hindi'

file           = datafolder_p + r'\1-Hinglish_SarcasmCSV.csv'
file_clean0     = datafolder_p + r"\2-Hinglish_Sarcasm_Clean0.csv"
file_clean     = datafolder_p + r"\2-Hinglish_Sarcasm_Clean.csv"
file_ft        = datafolder_p + r"\2-Hinglish_Sarcasm_Clean-fasttext.csv"
train_datafile = datafolder_p + r'\2-train.csv'
test_datafile  = datafolder_p + r'\2-test.csv'

#libraries required
import pandas as pd
import nltk
# import re
# import numpy as np

df =pd.read_csv( file_clean0 , sep='\t')
corpus = list(df['sentence_wo_emo'])
tokenized_corpus = [nltk.word_tokenize(doc) for doc in corpus]

def tranlit2dev(doc):
    from indicnlp.transliterate import acronym_transliterator
    from indic_transliteration.sanscript import transliterate 
    from indic_transliteration import sanscript
    import re

    ack_transliterator=acronym_transliterator.LatinToIndicAcronymTransliterator()

    words=re.split(r' ',doc) #break sentence on space
    sent = []
    for w in words:
        if w[0]!='#': #Don't convert hashtags
            w1 = re.findall(r'[A-z]+',w) ##is  word Latin word?
            if len(w1)>0: #Found english word
                w2=w1[0]
                pat = re.compile(r'([A-Z][a-z]+)') #English word with camel case. Break camel case into different words
                w2= pat.split(w2)
                w4=[]
                for w3 in w2:
                    if len(w3)>0:    
                        if w3.upper()==w3: # all letters are capital in the word
                            w4.append( ack_transliterator.transliterate(w3,lang='hi') )
                        else:
                            w3=w3.lower()
                            w4.append( transliterate(w3, sanscript.ITRANS, sanscript.DEVANAGARI) )
                    else:
                        pass
                    #print (w4)
                w5="_".join(w4)

            else:
                w5=w
        else:
            w5=w

        sent.append(w5)
    return " ".join(sent)
    
    
doc = 'हुर्रियत जेहादी यासिन मलिक Modi और आजाद Kashmir कश्मीर चिलाने वाले कश्मीर को बचाने आगे नहीं आएंगे #JammuKashmirFloods. #KashmirFloods'
doc ='सरिया कानून के हिसाब से अब WIFE IS MOTHER'
tranlit2dev(doc)

#re.split(r'(A-Z)+','ModiForIndia')

p = re.compile(r'([A-Z][a-z]+)')
x= p.split('Sad')
x

x

df = df[['label','sentence']]
#save file required in fasttext format
df.to_csv(file_ft,sep="\t", index=None, header=None)

# Aksharamukha

!pip install aksharamukha
#http://aksharamukha.appspot.com/documentation

from aksharamukha import transliterate

inputtxt = "News Hari Modi Ashok Narendra Venu Namtra Vinita"
transliterate.process( 'IAST', 'Devanagari',inputtxt, pre_options=['RemoveSchwaHindi'])

inputtxt = "हरी मोदी नरेन्द्र वेणु नम्रता विनिता"
transliterate.process( 'Devanagari','IAST', inputtxt, pre_options=['RemoveSchwaHindi'])

transliterate.process('HK', 'Tamil', 'bRhaspati gaMgA', False, post_options = ['TamilSubScript','TamilRemoveApostrophe'])

transliterate.process('Thai', 'Devanagari', 'พุทธัง สะระณัง คัจฉามิ', pre_options=['ThaiOrthography'])

transliterate.process('Devanagari', 'IAST', 'धर्म भारत की श्रमण परम्परा से निकला धर्म और दर्शन है', pre_options=['RemoveSchwaHindi'])

transliterate.process('Devanagari', 'Kannada', 'धर्म भारत की श्रमण परम्परा से निकला धर्म और दर्शन है', pre_options=['RemoveSchwaHindi'])

inputtxt="വെള്ളിയാഴ്ച, തീർത്തും അപ്രതീക്ഷിതമായിട്ടാണ് പ്രധാനമന്ത്രി നരേന്ദ്ര മോദി ലഡാക്കിൽ എത്തിയത്. സംയുക്ത സ..."
transliterate.process('Malayalam', 'Devanagari', inputtxt, pre_options=['RemoveSchwaHindi'])

# indic_nlp Translation

#https://nbviewer.jupyter.org/url/anoopkunchukuttan.github.io/indic_nlp_library/doc/indic_nlp_examples.ipynb

#https://github.com/anoopkunchukuttan/indic_nlp_library
#https://github.com/anoopkunchukuttan/indic_nlp_resources
        
# The path to the local git repo for Indic NLP library
INDIC_NLP_LIB_HOME=r'D:\18-DS\github\Project_NLP\indic_nlp_library'

# The path to the local git repo for Indic NLP Resources
INDIC_NLP_RESOURCES=r'D:\18-DS\github\Project_NLP\indic_nlp_resources'

import sys
sys.path.append(r'{}\src'.format(INDIC_NLP_LIB_HOME))

from indicnlp import common
common.set_resources_path(INDIC_NLP_RESOURCES)

from indicnlp import loader
loader.load()

### Sentence Tokentization

from indicnlp.tokenize import sentence_tokenize

indic_string="""तो क्या विश्व कप 2019 में मैच का बॉस टॉस है? यानी मैच में हार-जीत में \
टॉस की भूमिका अहम है? आप ऐसा सोच सकते हैं। विश्वकप के अपने-अपने पहले मैच में बुरी तरह हारने वाली एशिया की दो टीमों \
पाकिस्तान और श्रीलंका के कप्तान ने हालांकि अपने हार के पीछे टॉस की दलील तो नहीं दी, लेकिन यह जरूर कहा था कि वह एक अहम टॉस हार गए थे।"""
sentences=sentence_tokenize.sentence_split(indic_string, lang='hi')
for t in sentences:
    print(t)

from indicnlp.tokenize import indic_tokenize  

indic_string='सुनो, कुछ आवाज़ आ रही है। फोन?'

print('Input String: {}'.format(indic_string))
print('Tokens: ')
for t in indic_tokenize.trivial_tokenize(indic_string): 
    print(t)

from indicnlp.tokenize import indic_detokenize  
indic_string='" सुनो , कुछ आवाज़ आ रही है . " , उसने कहा । '

print('Input String: {}'.format(indic_string))
print('Detokenized String: {}'.format(indic_detokenize.trivial_detokenize(indic_string,lang='hi')))

### Script Conversion

from indicnlp.transliterate.unicode_transliterate import UnicodeIndicTransliterator
from indicnlp.transliterate.script_unifier import unicode_transliterate
input_text='राजस्थान मुझे बहुत पसंद है'
# input_text='രാജസ്ഥാന'
# input_text='රාජස්ථාන'
print(UnicodeIndicTransliterator.transliterate(input_text,"hi","pa"))
print(UnicodeIndicTransliterator.transliterate(input_text,"hi","ta"))
print(unicode_transliterate.UnicodeIndicTransliterator(input_text,"hi","ta"))

### Romanization

from indicnlp.transliterate.unicode_transliterate import ItransTransliterator

input_text='राजस्थान'
# input_text='ஆசிரியர்கள்'
lang='hi'

print(ItransTransliterator.to_itrans(input_text,lang))

### Indicization

from indicnlp.transliterate.unicode_transliterate import ItransTransliterator


# input_text='rajasthAna'
input_text='tanni'
lang='ta'
x=ItransTransliterator.from_itrans(input_text,lang)
print(x)
for y in x:
    print('{:x}'.format(ord(y)))

### Query Phonetic Properties

from indicnlp.script import  indic_scripts as isc

c='क'
lang='hi'
y=isc.get_phonetic_feature_vector(c,lang)
y

sorted(isc.PV_PROP_RANGES.items(),key=lambda x:x[1][0])

from indicnlp.langinfo import *

c='क'
lang='hi'

print('Is vowel?:  {}'.format(is_vowel(c,lang)))
print('Is consonant?:  {}'.format(is_consonant(c,lang)))
print('Is velar?:  {}'.format(is_velar(c,lang)))
print('Is palatal?:  {}'.format(is_palatal(c,lang)))
print('Is aspirated?:  {}'.format(is_aspirated(c,lang)))
print('Is unvoiced?:  {}'.format(is_unvoiced(c,lang)))
print('Is nasal?:  {}'.format(is_nasal(c,lang)))

from indicnlp.script import  indic_scripts as isc
from indicnlp.script import  phonetic_sim as psim

c1='क'
c2='ख'
c3='भ'
lang='hi'

print('Similarity between {} and {}'.format(c1,c2))
print(psim.cosine(
    isc.get_phonetic_feature_vector(c1,lang),
    isc.get_phonetic_feature_vector(c2,lang)
    ))

print()

print(u'Similarity between {} and {}'.format(c1,c3))
print(psim.cosine(
    isc.get_phonetic_feature_vector(c1,lang),
    isc.get_phonetic_feature_vector(c3,lang)
    ))

from indicnlp.script import  indic_scripts as isc
from indicnlp.script import  phonetic_sim as psim


slang='hi'
tlang='ml'
sim_mat=psim.create_similarity_matrix(psim.cosine,slang,tlang,normalize=False)

c1='क'
c2='ഖ'
print('Similarity between {} and {}'.format(c1,c2))
print(sim_mat[isc.get_offset(c1,slang),isc.get_offset(c2,tlang)])

slang='hi'
tlang='ml'
sim_mat=psim.create_similarity_matrix(psim.sim1,slang,tlang,normalize=True)

c1='क'
c2='ഖ'
print(u'Similarity between {} and {}'.format(c1,c2))
print(sim_mat[isc.get_offset(c1,slang),isc.get_offset(c2,tlang)])

### Lexical Similarity

from indicnlp.script import  indic_scripts as isc
from indicnlp.transliterate.unicode_transliterate import UnicodeIndicTransliterator

lang1_str='पिछले दिनों हम लोगों ने कई उत्सव मनाये. कल, हिन्दुस्तान भर में श्री कृष्ण जन्म-महोत्सव मनाया गया.'
lang2_str='વીતેલા દિવસોમાં આપણે કેટલાય ઉત્સવો ઉજવ્યા. હજી ગઇકાલે જ પૂરા હિંદુસ્તાનમાં શ્રીકૃષ્ણ જન્મોત્સવ ઉજવવામાં આવ્યો.'
lang1='hi'
lang2='gu'

lcsr, len1, len2 = isc.lcsr_indic(lang1_str,lang2_str,lang1,lang2)

print('{} string: {}'.format(lang1, lang1_str))
print('{} string: {}'.format(lang2, UnicodeIndicTransliterator.transliterate(lang2_str,lang2,lang1)))
print('Both strings are shown in Devanagari script using script conversion for readability.')
print('LCSR: {}'.format(lcsr))

from indicnlp.transliterate import acronym_transliterator

ack_transliterator=acronym_transliterator.LatinToIndicAcronymTransliterator()
ack_transliterator.transliterate('PMO',lang='hi')

### Orthographic Syllabification

from indicnlp.syllable import  syllabifier

w='जगदीशचंद्र'
lang='hi'

print(' '.join(syllabifier.orthographic_syllabify(w,lang)))

from indicnlp.morph import unsupervised_morph 
from indicnlp import common

analyzer=unsupervised_morph.UnsupervisedMorphAnalyzer('mr')

indic_string='आपल्या हिरड्यांच्या आणि दातांच्यामध्ये जीवाणू असतात .'

analyzes_tokens=analyzer.morph_analyze_document(indic_string.split(' '))

for w in analyzes_tokens: 
    print(w)

# <font color=red>Language Detection</font>

### Using fasttext

import fasttext
ftmodel_lang = fasttext.load_model(vectorfolder+'\lid.176.bin')

# print('\n',len(ftmodel_lang.words) )
# print('\n',ftmodel_lang.get_dimension() )
# #print(ftmodel_lang.get_labels())
# print('\n',ftmodel_lang.get_line("hari"))
# print('\n',ftmodel_lang.get_analogies('apple','orange','juice',10) )

# print ('\n', ftmodel_lang.get_subwords('Hari') )
# #print( ftmodel_lang.get_words())
# print ('\n',ftmodel_lang.get_sentence_vector('Hari Thapliyal'))
# print ('\n',ftmodel_lang.get_subword_id('apple') )
# print ('\n',ftmodel_lang.get_nearest_neighbors('apple') )
# ftmodel_lang.predict(['Thapliyal', 'हर'])
n=27
list(zip(tokenized_corpus[n],ftmodel_lang.predict(tokenized_corpus[n])[0],ftmodel_lang.predict(tokenized_corpus[n])[1]))

 ### Using Langdtect

#!pip install langdetect 
from langdetect import detect

def whatLang(char):
    try:
        return detect(char)
    except:
        return "?"

sentences = list(df.sentence)
emoticon_fld=[]
hashtag_fld=[]
newSents=[]
for sentence in sentences: #for sentence
    words = nltk.word_tokenize(sentence)
    
    emotions =[]
    hashtags =[]
    i=0
    for word in words:
        lng = whatLang(word)
        print (word,lng)
                

#     emotions =' '.join(emotions)
#     emotions = "".join( set(emotions) )
#     hashtags = ",".join(hashtags)
    
#     emoticon_fld.append(emotions)
#     hashtag_fld.append(hashtags)

# #extract unique emotions of sentence and put in separate field. Without altering original text
# df['emoticon']=emoticon_fld
# df['hashtags']=hashtag_fld
# #df['sentence'] = newSents

# Transliteration

 ### Using indic-transliteration

#!pip install indic-transliteration

import nltk
from indic_transliteration import sanscript 
from indic_transliteration.sanscript import transliterate 
  
# the text to be transliterated 
text = "Apa sabhii kaa yahaan svaagat hai."
#text = "ThankYouAmitShah"
#text ="Behaviour"
tkn = nltk.word_tokenize(text)

# for k in tkn:
#     #text = "News"
#     # printing the transliterated text 
#     #print(transliterate(text, sanscript.ITRANS, sanscript.DEVANAGARI)) 
#     print(detect(k),k, transliterate(k, sanscript.ITRANS, sanscript.DEVANAGARI)) 

print(transliterate(text, sanscript.ITRANS, sanscript.DEVANAGARI)) 

data = 'idam adbhutam'
data = "narendra modi"
data ="hari thapaliyala news"
transliterate(data, sanscript.ITRANS, sanscript.DEVANAGARI)

### Using indic-nlp-library

#!pip install indic-nlp-library

### inltk

!pip install torch==1.3.1+cpu -f 
#'https://download.pytorch.org/whl/torch_stable.html'
#'https://download.pytorch.org/whl/torch_stable.html'

#!pip install inltk

#!pip install torch==1.3.1+cpu -f 'https://download.pytorch.org/whl/torch_stable.html'
    
from inltk.inltk import setup
setup('hi')

# The languages supported are Hindi (hi), Punjabi (pa), Sanskrit (sa), Gujarati (gu), Kannada (kn), 
# Malayalam (ml), Nepali (ne), Odia (or), Marathi (mr), Bengali (bn), Tamil (ta), Urdu (ur), English (en).


#Tokenize
from inltk.inltk import tokenize
text ='#News तीन साल बाथरूम में बंद रही विवाहिता'
tokenize( text,'hi')

from inltk.inltk import identify_language, reset_language_identifying_models

reset_language_identifying_models() # only if you've updated iNLTK version
identify_language(text)

from inltk.inltk import get_embedding_vectors

vectors = get_embedding_vectors('भारत', 'hi')
vectors[0].shape

get_embedding_vectors('ਜਿਹਨਾਂ ਤੋਂ ਧਾਤਵੀ ਅਲੌਹ ਦਾ ਆਰਥਕ','pa')
vectors = get_embedding_vectors('ਜਿਹਨਾਂ ਤੋਂ ਧਾਤਵੀ ਅਲੌਹ ਦਾ ਆਰਥਕ','pa')
len(vectors)


### Using indicnlp

from indicnlp.transliterate.unicode_transliterate import UnicodeIndicTransliterator

# Input text "Today the weather is good. Sun is bright and there are no signs of rain. Hence we can play today."
input_text='आज मौसम अच्छा है। सूरज उज्ज्वल है और बारिश के कोई संकेत नहीं हैं। इसलिए हम आज खेल सकते हैं!'

# Transliterate from Hindi to Telugu
print(UnicodeIndicTransliterator.transliterate(input_text,"hi","pa"))

from indic_transliteration.sanscript import SchemeMap, SCHEMES, transliterate
from indic_transliteration import sanscript
transliterate(input_text, sanscript.HK, sanscript.ITRANS)

code('A')

from indic_transliteration import sanscript
from indic_transliteration.sanscript import SchemeMap, SCHEMES, transliterate

data = 'EVM'

print(transliterate(data, sanscript.HK, sanscript.DEVANAGARI))

help(UnicodeIndicTransliterator.transliterate())

from indicnlp.transliterate.unicode_transliterate import ItransTransliterator

input_text='आज मौसम अच्छा है। इसलिए हम आज खेल सकते हैं!'

# Transliterate Hindi to Roman
print(ItransTransliterator.to_itrans(input_text, 'ro'))

#help(ItransTransliterator)

### Stanfordnlp

#!pip install stanfordnlp

import stanfordnlp
stanfordnlp.download('hi')

import stanfordnlp

stanfordnlp.download('en')

help(stanfordnlp.Pipeline)

nlp = stanfordnlp.Pipeline(processors = "pos" )


import stanfordnlp

config = {
        'processors': 'tokenize,pos',
        'tokenize_pretokenized': True,
        'pos_model_path': './en_ewt_models/en_ewt_tagger.pt',
        'pos_pretrain_path': './en_ewt_models/en_ewt.pretrain.pt',
        'pos_batch_size': 1000
         }
nlp = stanfordnlp.Pipeline(**config)
# doc = nlp('Joe Smith lives in California .\nHe loves pizza .')
# print(doc.conll_file.conll_as_string())

pos_dict = {
'CC': 'coordinating conjunction','CD': 'cardinal digit','DT': 'determiner',
'EX': 'existential there (like: \"there is\" ... think of it like \"there exists\")',
'FW': 'foreign word','IN':  'preposition/subordinating conjunction','JJ': 'adjective \'big\'',
'JJR': 'adjective, comparative \'bigger\'','JJS': 'adjective, superlative \'biggest\'',
'LS': 'list marker 1)','MD': 'modal could, will','NN': 'noun, singular \'desk\'',
'NNS': 'noun plural \'desks\'','NNP': 'proper noun, singular \'Harrison\'',
'NNPS': 'proper noun, plural \'Americans\'','PDT': 'predeterminer \'all the kids\'',
'POS': 'possessive ending parent\'s','PRP': 'personal pronoun I, he, she',
'PRP$': 'possessive pronoun my, his, hers','RB': 'adverb very, silently,',
'RBR': 'adverb, comparative better','RBS': 'adverb, superlative best',
'RP': 'particle give up','TO': 'to go \'to\' the store.','UH': 'interjection errrrrrrrm',
'VB': 'verb, base form take','VBD': 'verb, past tense took',
'VBG': 'verb, gerund/present participle taking','VBN': 'verb, past participle taken',
'VBP': 'verb, sing. present, non-3d take','VBZ': 'verb, 3rd person sing. present takes',
'WDT': 'wh-determiner which','WP': 'wh-pronoun who, what','WP$': 'possessive wh-pronoun whose',
'WRB': 'wh-abverb where, when','QF' : 'quantifier, bahut, thoda, kam (Hindi)','VM' : 'main verb',
'PSP' : 'postposition, common in indian langs','DEM' : 'demonstrative, common in indian langs'
}

hindi_doc = nlp("""केंद्र की मोदी सरकार ने शुक्रवार को अपना अंतरिम बजट पेश किया. कार्यवाहक वित्त मंत्री पीयूष गोयल \
ने अपने बजट में किसान, मजदूर, करदाता, महिला वर्ग समेत हर किसी के लिए बंपर ऐलान किए. हालांकि, \
बजट के बाद भी टैक्स को लेकर काफी कन्फ्यूजन बना रहा. केंद्र सरकार के इस अंतरिम बजट क्या खास रहा \
और किसको क्या मिला, आसान भाषा में यहां समझें""")

#extract parts of speech
def extract_pos(doc):
    parsed_text = {'word':[], 'pos':[], 'exp':[]}
    for sent in doc.sentences:
        for wrd in sent.words:
            if wrd.pos in pos_dict.keys():
                pos_exp = pos_dict[wrd.pos]
            else:
                pos_exp = 'NA'
            parsed_text['word'].append(wrd.text)
            parsed_text['pos'].append(wrd.pos)
            parsed_text['exp'].append(pos_exp)
    #return a dataframe of pos and text
    return pd.DataFrame(parsed_text)

import stanfordnlp

MODELS_DIR = '.'
stanfordnlp.download('en', MODELS_DIR) # Download the English models
nlp = stanfordnlp.Pipeline(processors='tokenize,pos', models_dir=MODELS_DIR, treebank='hi_ewt', use_gpu=True, pos_batch_size=3000) # Build the pipeline, specify part-of-speech processor's batch size
doc = nlp("Barack Obama was born in Hawaii.") # Run the pipeline on input text
doc.sentences[0].print_tokens() # Look at the result

!pip install stanza

import stanza
stanza.download('en') # download English model
nlp = stanza.Pipeline('en') # initialize English neural pipeline
doc = nlp("Barack Obama was born in Hawaii.") # run annotation over a sentence

import stanza

import stanfordnlp

MODELS_DIR = '.'
stanfordnlp.download('en', MODELS_DIR) # Download the English models
nlp = stanfordnlp.Pipeline(processors='tokenize,pos', models_dir=MODELS_DIR, treebank='en_ewt', use_gpu=True, pos_batch_size=3000) 
# Build the pipeline, specify part-of-speech processor's batch size
doc = nlp("Barack Obama was born in Hawaii.") # Run the pipeline on input text
doc.sentences[0].print_tokens() # Look at the result

data ='Sad_Stories_Of_Twitter'

from indic_transliteration import xsanscript
#from indic_transliteration.xsanscript import  SchemeMap, SCHEMES, transliterate
from indic_transliteration.xsanscript import  transliterate
#data = 'नेहरू चाचा का सन्देश। जिसे पढकर आप को चाचा को जहन्नुम से धरती पर लाकर चार जूते मारने का मन करेगा 😡 👇 #Nehru #नेहरू #कांग्रेसी #CongressMuktBharat Please #रिट्वीट'
#data = "Vishal Tripathi"
#data = "WhyMediaIsSoAntiHindu"

#print(transliterate(data, xsanscript.DEVANAGARI, xsanscript.KANNADA))
print(transliterate(data, xsanscript.IAST, xsanscript.DEVANAGARI))
print(transliterate(data, xsanscript.DEVANAGARI, xsanscript.IAST))

data='हिरण, हिंदू, हिंदुस्तान_को_हिन्दूराष्ट्र_घोषित_करें, हल्ला_बोल, हर_हर_महादेव, सुप्रभात, \
सुना_है,iphone7, सीधासवाल, सबूत, शुभरात्रि, शत्रुघ्न,ठीक_है, व्यंग्य,कटाक्ष, व्यंग, विश्व_विख्यात,पप्पू, विकास_को_इंसाफ_कब, वावा, \
वसंतपंचमी, लौटकेराहुलघरकोआए,तपस्या, लोकतंत्र, लॉलीपॉप, लादेन,ये, लल्लनटॉप_सरकार, लम्बाई, राहुल_गाँधी,rahul_gandhi, \
रामदेव,भक्त, रानी_लक्ष्मीबाई, मोदी_फिरकी_ले_रहा_है, मोदी,केजरी, मोदी,2019Elections, मोदी, मुंबई, मायावती,शम्भुनाथ,हिंदुत्व, \
महादेव,शिव_, महादेव, महातूफान,उम्पुन, ममता_का_जंगलराज, मजदूर, भैरोंसिंह_शेखावत, भारतीय_सेना_हमें_आप_पर_गर्व_है, भारत_बंद, \
भारत_को_हिन्दुराष्ट्र_घोसित_करो, भाजपा_का_विजयरथ, ब्राह्मण_।, बेहद_दुखद_खबर, बेचारी_बहन_का_बेचारा_दल्ला_भाई, \
बेगानी_शादी_में_अब्दुल्ला_दीवाना, बालाकोट, बारिश, बलात्कारी_बापू, बधाई_बेशर्म_केजरीवाल, बटवारा,\
PulwamaAttack,PulwamaEncounter,Pulwama,PulwamaTerrorAttack,पुलवामा_घात,पुलवामा, फ्री,कूट, \
फिलस्तीन,नेतन्याहू, प्रधानमंत्री,राजमाता।, पूर्वांचल_विरोधी_केजरीवाल, पाकिस्तान, पप्पू_दिवस, पप्पू,राहुल_गांधी,राहुल,RahulGandhiInDubai, \
पपु_यादव, पद्मावत, पकिस्तान, नशा_मुक्ति_भारत_अभियान, नमो, धरना_भी, धन्य_है_भारत_भूमि, दोगली_कांग्रेस, दिल्ली_सर, \
दगाबाज_चीन, तीन,पाँच,निजी_अनुभव, तिब्बत,JunkOneChina, ड्रामेबाज_कांग्रेस, झूठ_लूट_की_सरकार, जोकर, जयहिंद,वंदेमातरम, \
जयश्रीराम, जय_हिंद_की_सेना, जय_श्रीराम,जय_बजरंगबली, जय_श्रीराम, जय_श्री_कृष्णा,सुप्रभात।, जन्मदिन_मुबारक_बादशाह, \
छत्तीसगढ़,माओवादियों,संपत्ति,कुर्क, चुनाव,अब, चीनी_चमचे_भाई_भाई, गौमाता_को_इंसाफ_दो, गृहमंत्री,नाराज_फूफाजी'

data='अँधेरे, अंगना, अकेला, अनिलकपूर, अब, अमित_शाह, अम्बेडकर, अयाश_आसाराम, अय्याश, अय्याश_आसाराम, अय्याशी, अरेस्ट_चंद्रशेख।, अरेस्ट_चंद्रशेखर_रावण, अरेस्ट_चन्द्रशेखर_रावण, अरेस्ट_पंकज_पूनिया, असली, आरक्षण, आरक्षण_भीख_है, आशा, आशुतोष, आसाराम, आसाराम_की_नाजायज_औलाद_हितेन्द्र, इंशाल्ह_covid_played_well, इटालियन_परिवार_तिजोरी_खोलो, उम्पुन, ऊंट, एकऔरएकग्यारह, एडीजीपी, ओणम, कंगना, कटाक्ष, कदम, कांग्रेस_मुक्त_भारत, कांग्रेसियों, कांग्रेसी, किरण, कुर्क, कूट, केजरी, केजरीवाल, कोरोना, कोरोनावायरस, खूनीDemonetisation, गांव_बासनी_हरिसिंह, गिरफ्तार_करो_ओवैसी_को, गुफा, गृहमंत्री, गौमाता_को_इंसाफ_दो, चमचे_मुल्ले, चीनी_चमचे_भाई_भाई, चुनाव, छत्तीसगढ़, जन्मदिन_मुबारक_बादशाह, जय_बजरंगबली, जय_श्री_कृष्णा, जय_श्री_महांकाल, जय_श्रीराम, जय_हिंद, जय_हिंद_की_सेना, जयश्रीराम, जयहिंद, जयहिंद।, जागो।, जातपात, जोकर, झूठ_लूट_की_सरकार, ठीक_है, ड्रामेबाज_कांग्रेस, तपस्या, तिब्बत, तीन, दगाबाज_चीन, दिल्ली_सर, दोगली_कांग्रेस, धन्य_है_भारत_भूमि, धरना_भी, नमो, नमो_अगेन, नमो_नमः, नववर्ष, नववर्षकीशुभकामनाएँ, नशा_मुक्ति_भारत_अभियान, नाराज_फूफाजी, निजी_अनुभव, नेतन्याहू, नेहरू, पकिस्तान, पथ, पद्मावत, पपु_यादव, पप्पू, पप्पू_दिवस, पाँच, पाकिस्तान, पुलवामा, पुलवामा_आतंकी_हमला, पुलवामा_घात, पुलवामा_हमला, पूर्वांचल_विरोधी_केजरीवाल, प्रधानमंत्री, फिलस्तीन, फ्री, बटवारा, बधाई_बेशर्म_केजरीवाल, बलात्कारी_बापू, बारिश, बालाकोट, बेगानी_शादी_में_अब्दुल्ला_दीवाना, बेचारी_बहन_का_बेचारा_दल्ला_भाई, बेहद_दुखद_खबर, ब्राह्मण_।, भक्त, भाजपा_का_विजयरथ, भारत_को_हिन्दुराष्ट्र_घोसित_करो, भारत_बंद, भारतीय_सेना_हमें_आप_पर_गर्व_है, भैरोंसिंह_शेखावत, मजदूर, मदरसों, ममता_का_जंगलराज, महा_चू, महातूफान, महादेव, महासमर, माओवादियों, मायावती, मुंबई, मोदी, मोदी_फिरकी_ले_रहा_है, मोदी_लाओ_देश_बचाओ, यूँही, ये, राजमाता।, रानी_लक्ष्मीबाई, रामदेव, राष्ट्रवाद, राहुल, राहुल_गाँधी, राहुल_गांधी, राहुलगांधी, रिट्वीट, लम्बाई, लल्लनटॉप_सरकार, लादेन, लॉलीपॉप, लोकतंत्र, लौटकेराहुलघरकोआए, वंदेमातरम, वसंतपंचमी, वावा, विकास_को_इंसाफ_कब, विश्व_विख्यात, विहान, व्यंग, व्यंग्य, शत्रुघ्न, शम्भुनाथ, शिव_, शुभरात्रि, संपत्ति, सच, सबूत, सीधासवाल, सुना_है, सुप्रभात, सुप्रभात।, हर_हर_महादेव, हल्ला_बोल, हिंदुत्व, हिंदुस्तान_को_हिन्दूराष्ट्र_घोषित_करें, हिंदू, हिन्दू_हिन्दू_भाई_भाई, हिरण'


# Expertiments Area

transliterate('News', sanscript.roman, sanscript.DEVANAGARI)
