==> Media+Publication/SDSHL/SDSHL/notebooks/2-experiments/05-Feature_Engineering.ipynb <==
#paths required
datafolder_p = r'D:\18-DS\github\SDSHL\data\processed'
datafolder_e = r'D:\18-DS\github\SDSHL\data\external'
datafolder_i = r'D:\18-DS\github\SDSHL\data\internal'
modelfolder  = r'D:\18-DS\data\models\POS-Tagger-Hindi\hi_hdtb_models'

filepath_data = datafolder_p + r'\2-Hinglish_Sarcasm_Clean'

filepath_pos = datafolder_p + r'\4-features_pos.csv'

import pandas as pd

df_em = pd.read_csv(datafolder_e + "\Emoticons.csv")
emoticonList= list(df_em.Browser)

#To start feature engieering start with pos features. Thease are already identified in a file.

df= pd.read_csv( filepath_pos, sep='\t', index_col='ID' )

df.head(3)

### Number of english words

df['eng_words'] = df['sentence_wo_emo'].str.findall(r'[A-z]+').apply(len)

### Extract emoticons & hashtag and create features

import nltk
sentences = list(df.sentence)
emoticon_fld=[]
emoticon_uniq_fld=[]
hashtag_fld=[]
newSents=[]
for sentence in sentences: #for sentence
    words = nltk.word_tokenize(sentence)
    
    emotions =[]
    hashtags =[]
    i=0
    for word in words:
        #Extract emotions from text
        if word in emoticonList:
            emotions.append(word)
                
        if word=="#":
            #if not isDigit( words[i+1]):  #if number then don't create hashtag of it.
            if not words[i+1].isdigit():  #if number then don't create hashtag of it.
                hashtags.append(words[i+1])
            
        i+=1     

    emotions =','.join(emotions)
    emotions_uniq = ",".join( set(emotions) )
    hashtags = ",".join(hashtags)
    
    emoticon_fld.append(emotions)
    emoticon_uniq_fld.append(emotions_uniq)
    hashtag_fld.append(hashtags)

#extract unique emotions of sentence and put in separate field. Without altering original text
df['emoticons']=emoticon_fld
df['emoticons_uniq'] = emoticon_uniq_fld
df['hashtags']=hashtag_fld
#df['sentence'] = newSents

## Count hashtags & emoticons

def list_len(x):
    if len(x):
        y = len(x.split(","))
    else:
        y=0
    return y

df['no_emotion'] = df['emoticons'].apply(list_len)
df['no_hashtag'] = df['hashtags'].apply(list_len)

# <font color=red> Extract More Features in this section </font>

# save the extracted features

#after all feature engineering remove all non numeric field and save the file for model processing.
y=df['label']
df=df.drop(columns=['sentence','sentence_wo_emo','pos','emoticons','emoticons_uniq','hashtags','label'])
df['label']=y

df.to_csv(datafolder_p + r"\4-Hinglish_Sarcasm_Clean_FE.csv", sep='\t')

df

# Experiments

# words=['👻💕','Hari','😝😷','prasad', '😷😬😂']
# sent=[]
# emotions=[]

# print (emotions)



a[-1]

nltk.

regex = re.compile(r'\d+(.*?)(?:\u263a|\U0001f645)')

regex = re.compile(r'\d+(.*?)(?:\u263a|\U0001f645)')

regex

print(("✍").encode("unicode_escape"))
#print(("🇮🇳").encode("unicode_escape"))
#print(("क्या").encode("unicode_escape"))

!pip install langdetect 
from langdetect import detect

def isText(char):
    try:
        lng = detect(char)
        return True
    except:
        return False

a='✍🇮🇳क्या'
isText(a[0])

U+1F1EE,U+1F1F3

txt="सब गए ना 2019 में 👇 कोई छूटा तो नहीं 2018 में 😝 😅 😂 😆 #2018 #2019 #HappyNewYear2019 #HappyNewYear #नववर्षकीशुभकामनाएँ #नववर्ष"


words = nltk.word_tokenize(txt)
    
emotions =[]
hashtags =[]
i=0
for word in words:



    i+=1
hashtags

len(words)

def isNumeric(word):
    try:
        num = int(word)
        return True
    except:
        return False
