==> Media+Publication/SDSHL/SDSHL/notebooks/2-experiments/06-Dataset_Splitting.ipynb <==
#paths required
datafolder_p = r'D:\18-DS\github\SDSHL\data\processed'
datafolder_e = r'D:\18-DS\github\SDSHL\data\external'
datafolder_i = r'D:\18-DS\github\SDSHL\data\internal'
modelfolder  = r'D:\18-DS\data\models'

vectorfolder_bert = modelfolder + r'\BERT'
vectorfolder_ft   = modelfolder + r'\fasttext_wiki.hi'
vectorfolder_standford = modelfolder + r'\POS-Tagger-Hindi'

train_datafile =datafolder_p+r'\2-train.csv'
test_datafile  =datafolder_p+r'\2-test.csv'

file_clean =datafolder_p+ r'\2-Hinglish_Sarcasm_Clean.csv'

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split 

#split dataset for train & test

df =pd.read_csv(file_clean, sep="\t", index_col='ID')
y=df.label
#df.index = df['ID']

#Keep 10% for Testing
train,test=train_test_split(df, stratify=y, test_size=0.1,random_state=1000)
print (train.label.sum()/len(train), test.label.sum()/len(test))

#create train & test file for non-fasttest modeling
train = train[['label','sentence','sentence_wo_emo']]
test  = test[['label','sentence','sentence_wo_emo']]
train.to_csv(datafolder_p+ r'\2-train.csv', sep="\t")
test.to_csv (datafolder_p+ r'\2-test.csv', sep="\t")


#create train & test file for fasttest modeling
train = train[['label','sentence']]
test  = test[['label','sentence']]
test['label']= test['label'].apply(lambda x: '__label__'+str(x))
train['label']= train['label'].apply(lambda x: '__label__'+str(x))
train.to_csv(datafolder_p+ r'\3-train_ft.csv', sep="\t",  header=None, index=None)
test.to_csv (datafolder_p+ r'\3-test_ft.csv', sep="\t",  header=None, index=None)


#create fasttext full file for creating fasttext word embedding for the complete file
df  = df[['label','sentence']]
df['label']= df['label'].apply(lambda x: '__label__'+str(x))
df.to_csv(datafolder_p+ r'\3-Hinglish_Sarcasm_Clean_ft.csv', sep="\t", header=None, index=None)



# Expriment Area

import re

xx = "#AreYouSeriousRahul का सवाल पूछने वालों की SERIOUSNESS पर ही सवाल है."
#xx = 'हम तो इस देश की अर्थ व्यवस्था का अर्थ समझना चाह रहे थे Manmohan . तुमने तो अर्थ व्यस्था की अर्थी ही निकाल दी .'
r1 = re.split(r'[A-z]+',xx)

print(r1)

#list( df['sentence'].str.extract(r'[A-z]+') )

df_FE = pd.read_csv(datafolder_p + r'\4-features_pos.csv', sep='\t', index_col='ID')

df_FE
