==> Media+Publication/SDSHL/SDSHL/notebooks/2-experiments/11.1-embedding_tfid.ipynb <==
#paths required
datafolder_p = r'D:\18-DS\github\SDSHL\data\processed'
datafolder_e = r'D:\18-DS\github\SDSHL\data\external'
datafolder_i = r'D:\18-DS\github\SDSHL\data\internal'
modelfolder  = r'D:\18-DS\data\models'


modelfolder_w2v = modelfolder + r'\word2vec'

filepath_train  = datafolder_p+ r'\2-train.csv'
filepath_fulldata = datafolder_p+ r'\2-Hinglish_Sarcasm_Clean.csv'

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
import pandas as pd
from sklearn.decomposition import PCA

tfidf_vectorizer = TfidfVectorizer(max_features=300, ngram_range=(1,2) )

#emb = tfidf_vectorizer.fit_transform(df['sentence'])

def tfidf_vec(data):
    tfidf_vectorizer = TfidfVectorizer(max_features=300, ngram_range=(1,2) )

    emb = tfidf_vectorizer.fit_transform(data)

    return emb, tfidf_vectorizer

def bow_vec(data):
    count_vectorizer = CountVectorizer(max_features=300, ngram_range=(1,2) )

    emb = count_vectorizer.fit_transform(data)

    return emb, count_vectorizer


df      = pd.read_csv(filepath_train, sep='\t', index_col='ID')
df_full = pd.read_csv(filepath_fulldata, sep='\t', index_col='ID')

list_corpus      = df["sentence"].tolist()
list_corpus_full = df_full["sentence"].tolist()

## TFIDF Vector

embed_tfidf, vec_tfidf   = tfidf_vec(list_corpus_full)

embed_tfidf  = embed_tfidf.todense()

pca = PCA(n_components=15)
embed_tfidf  = pca.fit_transform(embed_tfidf)

embed_tfidf          = pd.DataFrame(embed_tfidf, index=df_full.index)
embed_tfidf['label'] = df_full['label']

embed_tfidf.shape

embed_tfidf.to_csv(datafolder_p+ r"\embedding_tfidf.csv")

## BOW Vector

embed_bow, vec_bow = bow_vec(list_corpus_full)
embed_bow = embed_bow.todense()

pca = PCA(n_components=15)
embed_bow = pca.fit_transform(embed_bow)

embed_bow = pd.DataFrame(embed_bow, index=df_full.index)
embed_bow['label'] = df_full['label']

embed_bow.shape

embed_bow.to_csv(datafolder_p+ r"\embedding_bow.csv")
