==> Media+Publication/SDSHL/SDSHL/notebooks/2-experiments/11.3-embedding_fastText.ipynb <==
#https://indicnlp.ai4bharat.org/indicft/


#paths required
datafolder_p = r'D:\18-DS\github\SDSHL\data\processed'
datafolder_e = r'D:\18-DS\github\SDSHL\data\external'
datafolder_i = r'D:\18-DS\github\SDSHL\data\internal'
modelfolder  = r'D:\18-DS\data\models'
resultsfolder = r'D:\18-DS\github\SDSHL\data\results'

modelfolder_ft       = modelfolder + r'\fasttext_wiki.hi'
modelfolder_ft_ind   = modelfolder + r'\fasttext_indicnlp.hi'
modelfolder_ft_local   = modelfolder + r'\fasttext_local'


#fasttax pretrained Hindi binary/text format vectors 
#Common Crawl and Wikipedia 
cc300 = r'\cc.hi.300.bin'
cc100 = r'\cc.hi.100.bin'

#wiki
wiki100    = r'\wiki.hi.100.bin'
wiki300    = r'\wiki.hi.300.bin'
wiki300_vec= r'\wiki.hi.300.vec'

#indicnlp
indicnlp300     = r'\indicnlp.ft.hi.300.bin'
indicnlp300_vec = r'\indicnlp.ft.hi.300.vec'

# FastText Model donot need dataset but filename in a specific fasttext format for training
train_file = datafolder_p+ r'\3-train_ft.csv'
test_file = datafolder_p+ r'\3-test_ft.csv'

# to predict the result with ID we need this file. Because fasttext format doesnot have ID
#train_file1 = datafolder_p+ r'\2-train.csv'
test_file1 = datafolder_p+ r'\2-test.csv'

#This file is needed to create fasttext sentence embedding of full dataset 
filepath_fulldata = datafolder_p+ r'\2-Hinglish_Sarcasm_Clean.csv'

# import pandas as pd

# file = datafolder_p + r'\2-Hinglish_Sarcasm_Clean.csv'
# df = pd.read_csv(file, sep='\t', index_col="ID")
# idx = list(df.index)

# file =  datafolder_p+ r'\2-train.csv' 
# df = pd.read_csv(file, sep='\t', index_col="ID")
# idx_train = df.index

# file =  datafolder_p+ r'\2-test.csv'
# df = pd.read_csv(file, sep='\t', index_col="ID")
# idx_test = df.index

# df = pd.read_csv(file_data, sep='\t', header=None)
# df['ID']=idx
# df=df.set_index('ID')

#!pip install keras
#!pip install pytorch
#!pip install tensorflow

!pip install fasttext

!pip install nltk

#from __future__ import print_function

import pandas as pd
import nltk
import numpy as np
import re
import sklearn.metrics as metrics

#!pip uninstall fasttext
import fasttext
from fasttext import load_model

#nltk.download('gutenberg')
#nltk.download('genesis')
#nltk.download('inaugural')
#nltk.download('brown')
#nltk.download('webtext')
#nltk.download('nps_chat')
nltk.download('treebank')


#C:\Users\admin\AppData\Roaming\nltk_data.
from nltk.book import nps_chat

df =pd.read_csv(train_file, index_col=None, header=None, sep='\t')
corpus = list(df[1])
tokenized_corpus = [nltk.word_tokenize(doc) for doc in corpus]
df['sentence_tkn'] = tokenized_corpus

#This is required to create word embedding of full dataset
df_full =pd.read_csv(filepath_fulldata, usecols=['ID','label','sentence'], sep="\t", index_col="ID")
corpus_full = list(df_full['sentence'])
tokenized_corpus_full = [nltk.word_tokenize(doc) for doc in corpus_full]
df_full['sentence_tkn'] = tokenized_corpus_full

# #Load 2-train data in memory. This file is required for lookup and creating wordvec for each sentence.
# df =pd.read_csv(train_file1, usecols=['ID','label','sentence'], sep="\t", index_col="ID")
# corpus = list(df['sentence'])
# tokenized_corpus = [nltk.word_tokenize(doc) for doc in corpus]
# df['sentence_tkn'] = tokenized_corpus

df_test=pd.read_csv(test_file1, usecols=['ID','label','sentence'], sep="\t", index_col="ID")
y_test =np.array(df_test['label'])
X_test=df_test['sentence']

# Some Common Functions

# Language Identification using fasttext

ftmodel_lang = fasttext.load_model(vectorfolder+'\lid.176.bin')

# print('\n',len(ftmodel_lang.words) )
# print('\n',ftmodel_lang.get_dimension() )
# #print(ftmodel_lang.get_labels())
# print('\n',ftmodel_lang.get_line("hari"))
# print('\n',ftmodel_lang.get_analogies('apple','orange','juice',10) )

# print ('\n', ftmodel_lang.get_subwords('Hari') )
# #print( ftmodel_lang.get_words())
# print ('\n',ftmodel_lang.get_sentence_vector('Hari Thapliyal'))
# print ('\n',ftmodel_lang.get_subword_id('apple') )
# print ('\n',ftmodel_lang.get_nearest_neighbors('apple') )
# ftmodel_lang.predict(['Thapliyal', 'हर'])
n=27
list(zip(tokenized_corpus[n],ftmodel_lang.predict(tokenized_corpus[n])[0],ftmodel_lang.predict(tokenized_corpus[n])[1]))

# Get Word Vector Without Pre-Training

# Set values for various parameters
feature_size = 15    # Word vector dimensionality  
window_context = 20  # Context window size                                                                                    
min_word_count = 1   # Minimum word count                        
sample = 1e-3        # Downsample setting for frequent words
sg = 1               # skip-gram model

#window_length = 200 # The amount of words we look at per example. Experiment with this.

def average_word_vectors(words, model, vocabulary, num_features):
    feature_vector = np.zeros((num_features,),dtype="float64")
    nwords = 0.
    for word in words:
        if word in vocabulary: 
            nwords = nwords + 1.
            feature_vector = np.add(feature_vector, model.wv[word])
    if nwords:
        feature_vector = np.divide(feature_vector, nwords)
    return feature_vector

def averaged_word_vectorizer(corpus, model, num_features):
    vocabulary = set(model.wv.index2word)
    features = [average_word_vectors(tokenized_sentence, model, vocabulary, num_features)
                    for tokenized_sentence in corpus]
    return np.array(features)

##THIS STEP TAKES VERY LONG TIME

from gensim.models.fasttext import FastText
ft_model = FastText(tokenized_corpus, size=feature_size, 
                          window=window_context, min_count = min_word_count,
                          sg=sg, sample=sample, iter=5000)

#save the model
ft_model.save(modelfolder_ft_local + r'\model_fasttext_local.bin')

#ft_tnseEmbedded, ft_words, ft_wvs, ft_labels = getTNSE_Embedded(ft_model)

# get document level embeddings
ft_doc_features = averaged_word_vectorizer(corpus=tokenized_corpus_full, model=ft_model,
                                             num_features=feature_size)

ft_features = pd.DataFrame(ft_doc_features, index=df_full.index)
ft_features['label']=df_full.label
ft_features.to_csv(datafolder_p + r"\embedding_ft_local.csv")
ft_features

# Using Pretrained Vector

### How to compress large size pretrained binary model

#load cc.hi.300.bin 
ft300 = fasttext.load_model(modelfolder_ft+ r'\cc.hi.300.bin')
ft300.get_dimension()

#reduce dimension from 300 to 100
import fasttext.util
fasttext.util.reduce_model(ft300, 100)
ft300.get_dimension()

#save the model with reduced dimension
ft300.save_model(vectorfolder+'\cc.hi.100.bin')

### Method 1: Load a pretrained model in memory

#ftmodel_cc300 = fasttext.load_model(modelfolder_ft+cc300)

print('Model file name:'+modelfolder_ft+cc300)
#print (ftmodel_cc300.get_dimension(), len(ftmodel_cc300.words))

ftmodel_cc100 = fasttext.load_model(modelfolder_ft+cc100)

print('file name:'+modelfolder_ft+cc100)
#print (ftmodel_cc100.get_dimension(), len(ftmodel_cc100.words))

ftmodel_indnlp300 = fasttext.load_model(modelfolder_ft_ind+indicnlp300)

print('Model file name:'+modelfolder_ft_ind+indicnlp300)
#print (ftmodel_indnlp300.get_dimension(), len(ftmodel_indnlp300.words))

### Method 3: Read details of Pretrained Vector of Text format

#fasttest wiki indicnlp vector

import io
def load_vectors(fname):
    fin = io.open(fname, 'r', encoding='utf-8', newline='\n', errors='ignore')
    n, d = map(int, fin.readline().split())
    data = {}
    for line in fin:
        tokens = line.rstrip().split(' ')
        data[tokens[0]] = map(float, tokens[1:])
    return data

dict_ftvec=load_vectors(modelfolder_ft+wiki300_vec)
print ( len(dict_ftvec))

### Method 4: How to create model from pretrained-vector

### <font color=red>wiki300_vec<font>

#used pretrained vector to get the word embedding. This model can also be used for prediction
#df should be in fastText format
print (train_file)
ftmodel_wiki300_vec = fasttext.train_supervised(train_file,dim=300, \
                                     lr=0.5, epoch=25, wordNgrams=2, bucket=200000,\
                                     pretrainedVectors=modelfolder_ft + wiki300_vec)
#loss='ova'

len(ftmodel_wiki300_vec.words)

print ( 'wiki300_vec has 158016 words - Method 3')
print ( 'ftmodel_wiki300_vec - a finetuned model from wiki300_vec vector has  160659 word - Method 4')
print ( 'ftmodel_vec300 model is created from existing vector')
#It means 160659-158016 are the new words in the fine tuned model as done above.

### <font color=red>indicnlp300_vec<font>

#used pretrained vector to get the word embedding. This model can also be used for prediction
#df should be in fastText format

ftmodel_indicnlp300_vec = fasttext.train_supervised(train_file,dim=300, \
                                     lr=0.5, epoch=25, wordNgrams=2, bucket=200000,\
                                     pretrainedVectors=modelfolder_ft_ind + indicnlp300_vec)
#loss='ova'

len(ftmodel_indicnlp300_vec.words)

####  \****Indicnlp based model has more words than wiki based model\****

# Performance of wiki_pretrained model

prediction={}

#### ftmodel_wiki300_vec

ftmodel_wiki300_vec.test(train_file)

def pred_for_test(model):
    y_pred=[]
    for i in range(len(X_test)):
        p=model.predict(X_test.iloc[i])
        x1 = int(p[0][-1][-1])
        x2 = p[1][0]
        if x1==0: x2= 1-x2
        y_pred.append( x2 )
    return y_pred    


def print_matrics(y_pred):
    threshold=0.5
    y_pred1=[]
    for i in y_pred :
        if i>threshold:
            y_pred1.append(1)
        else:
            y_pred1.append(0)

    acc = np.round( metrics.accuracy_score(y_test,y_pred1), 2)
    recall = np.round( metrics.recall_score(y_test,y_pred1), 2)
    precision = np.round( metrics.precision_score(y_test,y_pred1), 2)
    f1 = np.round( metrics.f1_score(y_test,y_pred1), 2)
    roc = np.round( metrics.roc_auc_score(y_test,y_pred1), 2)

    print("Accuracy : ", acc )
    print("Recall   : ", recall )
    print("Precision: ", precision )
    print("F1       : ", f1 )
    print("ROC      : ", roc )

y_pred = pred_for_test(ftmodel_wiki300_vec)
prediction['fasttext_wiki_pretrained'] = y_pred
print_matrics(y_pred)

#### ftmodel_indicnlp300_vec

ftmodel_indicnlp300_vec.test(train_file)

ftmodel_indicnlp300_vec.test(test_file)

y_pred = pred_for_test(ftmodel_indicnlp300_vec)
prediction['fasttext_indicnlp_pretrained'] = y_pred
print_matrics(y_pred)

df_prediction= pd.DataFrame(prediction, index=df_test.index)
df_prediction.to_csv(resultsfolder+ r'\model_predictions_fasttext.csv')
df_prediction

# Convert Sentences to Vector & save Embedding

#Use ftmodel_cc100 to generate senetence vector. This vector later on can be \
#used in classical ML models for classification.

def df_to_vec(n_features,df,model ):

    sent_all_vec = np.zeros((len(df), n_features), dtype='float32')
    i=0
    for row in np.array(df):
        sent_vec=np.zeros(n_features)
        j=0
        for word in row:
            sent_vec=sent_vec+ model.get_word_vector(word).astype('float32') 
            j+=1
        sent_vec = sent_vec/j
        sent_all_vec[i]=sent_vec
        i+=1
    return sent_all_vec

#### ftmodel_wiki300_vec

#convert sentence to vector using pretrained fasttext
df1 = pd.DataFrame( df_to_vec(300, df_full['sentence_tkn'], ftmodel_wiki300_vec), index=df_full.index)
df1['label'] = df_full['label']

df1.to_csv(datafolder_p+r'\embedding_ft300_wiki_pretrained.csv')

#### ftmodel_indicnlp300_vec

#convert sentence to vector using pretrained fasttext
df1 = pd.DataFrame( df_to_vec(300, df_full['sentence_tkn'], ftmodel_indicnlp300_vec), index=df_full.index)
df1['label'] = df_full['label']

df1.to_csv(datafolder_p+r'\embedding_ft300_indicnlp_pretrained.csv')

df1

# Unsupervised Learning

### Method 5: How to cerate a new model using our corpus. This can be used as a pretrained model for other NLP tasks

import fasttext
# Skipgram model :
ftmodel_usup = fasttext.train_unsupervised(train_datafile, model='skipgram', dim=300, minn=2, maxn=5,\
                                     lr=0.5, epoch=5, wordNgrams=2, bucket=200000)

# or, cbow model :
#model = fasttext.train_unsupervised('data.txt', model='cbow')

print( len(model.words)) 
#print (model.words)
ftmodel_usup.save_model("sdsms.bin")

#load the saved & trained model.
ftmodel_usup = fasttext.load_model("sdsms.bin")

print(ftmodel_usup.get_dimension())
print(ftmodel_usup.get_words())

ft_doc_features = averaged_word_vectorizer(corpus=tokenized_corpus, model=ftmodel_usup,
                                             num_features=feature_size)

ftmodel_vec300.predict('हिंसा प्रति हिंसा को रोकना दुरुह कार्य है.')

# <font color=red> Experiments

# import io
# vecfile_name= vectorfolder + '\wiki.hi.vec'

# fin = io.open(vecfile_name, 'r', encoding='utf-8', newline='\n', errors='ignore')
# data = {}
# for line in fin:
#     tokens = line.rstrip().split(' ')
#     data[tokens[0]] = map(float, tokens[1:])

# for a in data["माता"]:
#     print (a)

import fasttext
#help(fasttext.FastText)

from fasttext import FastText

#help(FastText)

import fasttext


# test.get_line('I am hari')

# test.get_sentence_vector('Hari Thapliyal')

#test.get_input_vector(1000000)

#test.get_subwords('Hari Thapliyal')

test.get_label_id('orange')


# print('\n',len(test.words) )
# print('\n',test.get_dimension() )
# #print(test.get_labels())
# print('\n',test.get_line("hari"))
# print('\n',test.get_analogies('apple','orange','juice',10) )

# print ('\n', test.get_subwords('Hari') )
# #print( test.get_words())
# print ('\n',test.get_sentence_vector('Hari Thapliyal'))
# print ('\n',test.get_subword_id('apple') )
# print ('\n',test.get_nearest_neighbors('apple') )
# test.predict(['Thapliyal', 'हर'])
n=1000
list(zip(tokenized_corpus[n],test.predict(tokenized_corpus[n])[0],test.predict(tokenized_corpus[n])[1]))

import numpy as np
a = np.array(tokenized_corpus)
a.reshape()

b=[]
for c  in a:
    b.append( " ".join(c))
    

a=set("".join(b).split(" "))

print(a)

len(tokenized_corpus)
tokenized_corpus[0]

b=np.array(a)
b.shape
c= set( list(b))

#fasttext.FastText(tokenized_corpus)

test= supervised(tokenized_corpus, size=feature_size, 
                              window=window_context, min_count = min_word_count,
                              sg=sg, sample=sample, iter=5000)

datetime.sys()
