==> Media+Publication/SDSHL/SDSHL/notebooks/2-experiments/11.5-embedding_RoBERTa.ipynb <==
!pip install RoBERTa
from transformers import RobertaConfig, RobertaModel, RobertaTokenizer, RobertaForMaskedLM

#model = BertModel.from_pretrained('bert-base-uncased') 
model2  = RobertaModel.from_pretrained('roberta-base')

tokenizer = RobertaTokenizer.from_pretrained('roberta-base')
tokenizer("Hello world")['input_ids']

model3 = RobertaForMaskedLM.from_pretrained('roberta-base')

input_ids = tokenizer("Hello, my dog is cute", return_tensors="pt")["input_ids"]
input_ids = tokenizer("मन्दिरो के सोने पर नजर गढाने वालो देश को लूटने", return_tensors="pt")["input_ids"]

outputs = model3(input_ids, labels=input_ids)
loss, prediction_scores = outputs[:2]
