==> Media+Publication/SDSHL/SDSHL/notebooks/2-experiments/11.6-embedding_GloVe_English.ipynb <==
# <font color=red> I have only GloVe vector for English Lanuage so cannot be used for Hinglish </font>

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split 

from keras.preprocessing.text import Tokenizer
from keras.preprocessing.sequence import pad_sequences

#paths required
datafolder_p = r'D:\18-DS\github\SDSHL\data\processed'
datafolder_e = r'D:\18-DS\github\SDSHL\data\external'
datafolder_i = r'D:\18-DS\github\SDSHL\data\internal'

vectorfolder_glove = r'D:\18-DS\data\Datasets_nlp\glove6b50dtxt\glove.6B.50d.txt'

df =pd.read_csv(datafolder_p+ r'\train.csv', sep='\t')

sentences.shape

sentences = df['sentence']
y = df['label']

sentences_train, sentences_val, y_train,y_val = train_test_split(sentences, y, 
                                                                 stratify=y, test_size=0.2, random_state=1000)

tokenizer = Tokenizer(num_words=5000, oov_token=1)
tokenizer.fit_on_texts(sentences_train)


X_train = tokenizer.texts_to_sequences(sentences_train)
X_val = tokenizer.texts_to_sequences(sentences_val)
# Adding 1 because of  reserved 0 index
vocab_size = len(tokenizer.word_index) + 1                          

maxlen = 100

X_train = pad_sequences(X_train, padding='post', maxlen=maxlen)
X_val = pad_sequences(X_val, padding='post', maxlen=maxlen)

print( pd.Series(y_val).sum()/len(pd.Series(y_val)), pd.Series(y_train).sum()/len(pd.Series(y_train)))

def create_embedding_matrix(vectorfile, word_index, embedding_dim):
    vocab_size = len(word_index) + 1  
    # Adding again 1 because of reserved 0 index
    embedding_matrix = np.zeros((vocab_size, embedding_dim))

    with open(vectorfile, encoding="utf8") as f:
        for line in f:
            word, *vector = line.split()
            if word in word_index:
                idx = word_index[word] 
                embedding_matrix[idx] = np.array(vector, dtype=np.float32)[:embedding_dim]
            else:
                idx=0

    return embedding_matrix

embedding_dim = 50
embedding_matrix = create_embedding_matrix( vectorfolder_glove,
                                            tokenizer.word_index,  
                                            embedding_dim)

embedding_matrix
