==> Media+Publication/SDSHL/SDSHL/notebooks/2-experiments/11.7-embedding-gpt2.ipynb <==
https://huggingface.co/surajp/gpt2-hindi

from transformers import AutoTokenizer, AutoModelWithLMHead

tokenizer = AutoTokenizer.from_pretrained("surajp/gpt2-hindi",)

model = AutoModelWithLMHead.from_pretrained("surajp/gpt2-hindi" )


!pip list

import tensorflow

tensorflow.__version__
