==> Media+Publication/SDSHL/SDSHL/notebooks/2-experiments/11.8-embedding_cltk.ipynb <==
from cltk.tokenize.sentence import TokenizeSentence

from cltk.corpus.marathi.alphabet import DIGITS as marathi_digits
import os
root = os.path.expanduser('~')
hindi_corpus = os.path.join(root,'cltk_data/hindi/text/hindi_text_ltrc')
hindi_text_path = os.path.join(hindi_corpus, 'miscellaneous/gandhi/main.txt')
#hindi_text_path = os.path.join(hindi_corpus, 'Kabeera/main.txt')
hindi_text = open(hindi_text_path,'r', encoding="utf-8").read()
tokenizer = TokenizeSentence('hindi')
hindi_text_tokenize = tokenizer.tokenize(hindi_text)
print(hindi_text_tokenize[0:100])
