==> Media+Publication/SDSHL/SDSHL/notebooks/2-experiments/11.9-embedding_ELMo.ipynb <==
# Embedding for Language Modeles

import pandas as pd

datafolder = r'../../data/processed/'
df= pd.read_csv(datafolder+"Hinglish_SarcasmCSV.csv")
sentences = list(df.sentence)

#!pip install pytorch

!pip install ELMoForManyLangs

from elmoformanylangs import Embedder

def sents2elmo(sents, output_layer=-1):
class Embedder(model_dir='/path/to/your/model/', batch_size=64):

modelpath = r'D:\18-DS\data\Datasets_nlp\ELMo-Hindi'
#/token_embedder.pkl
e = Embedder(modelpath)


#e = Embedder('/path/to/your/model/')

# sents = [['今', '天', '天氣', '真', '好', '阿'],
# ['潮水', '退', '了', '就', '知道', '誰', '沒', '穿', '褲子'],
#         ['भैया','लोगों','को','कोन','समझाए','तुमने']]

#sets = [['भैया','लोगों','को','कोन','समझाए','तुमने'],['शांति','और','दोस्ती','की','मे']]
# the list of lists which store the sentences 
# after segment if necessary.

sents = sentences
emb= e.sents2elmo(sents)
# will return a list of numpy arrays 
# each with the shape=(seq_len, embedding_size)

# Conclusion it is not working for our dataset
